# api/openapi.yml
openapi: 3.0.0
paths:
...
  /profile:
    get:
      operationId: ProfileController_getProfile
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileSummaryDto'
      security:
        - vaultfolio_session: []
      summary: Get the caller's own profile.
      tags: &ref_8
        - profile
  /profile/display-name:
    patch:
      operationId: ProfileController_updateDisplayName
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDisplayNameRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileSummaryDto'
        '400':
          description: Display name must be 1–100 characters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileErrorResponseDto'
      security:
        - vaultfolio_session: []
      summary: Update the caller's display name.
      tags: *ref_8
  /profile/email-language:
    patch:
      operationId: ProfileController_updateEmailLanguage
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmailLanguageRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileSummaryDto'
        '400':
          description: Email language must be a supported language or unset.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileErrorResponseDto'
      security:
        - vaultfolio_session: []
      summary: Update the caller's notification email language.
      tags: *ref_8
  /profile/email-change:
    post:
      operationId: ProfileController_requestEmailChange
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestEmailChangeRequestDto'
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  pendingEmail:
                    type: string
        '409':
          description: This email can't be used right now.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileErrorResponseDto'
        '502':
          description: Request saved, but the email could not be sent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileErrorResponseDto'
      security:
        - vaultfolio_session: []
      summary: Request an email change, sending a confirmation link to the new address.
      tags: *ref_8
  /profile/email-change/cancel:
    post:
      operationId: ProfileController_cancelEmailChange
      parameters: []
      responses:
        '204':
          description: Cancelled.
      security:
        - vaultfolio_session: []
      summary: Cancel a pending email change.
      tags: *ref_8
  /profile/email-change/token/{token}:
    get:
      operationId: ProfileController_lookupEmailChangeToken
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  newEmail:
                    type: string
        '410':
          description: This link is no longer valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileErrorResponseDto'
      summary: Look up a pending email change's new address by its token.
      tags: *ref_8
  /profile/email-change/token/{token}/confirm:
    post:
      operationId: ProfileController_confirmEmailChange
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
        '410':
          description: This link is no longer valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileErrorResponseDto'
      summary: Confirm a pending email change via its emailed token.
      tags: *ref_8
  /profile/password:
    post:
      operationId: ProfileController_changePassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  changed:
                    type: boolean
                    enum:
                      - true
        '400':
          description: Password must be between 8 and 200 characters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileErrorResponseDto'
        '401':
          description: Current password is incorrect.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileErrorResponseDto'
      security:
        - vaultfolio_session: []
      summary: Change the caller's password.
      tags: *ref_8
  /profile/forgot-password:
    post:
      description: Requires a valid Cloudflare Turnstile token (see `turnstileToken`).
      operationId: ProfileController_forgotPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequestDto'
      responses:
        '200':
          description: >-
            Always returns { accepted: true } regardless of whether the email
            exists.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted:
                    type: boolean
                    enum:
                      - true
      summary: Request a password-reset email.
      tags: *ref_8
  /profile/reset-password/token/{token}:
    get:
      operationId: ProfileController_lookupResetToken
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    enum:
                      - true
                  displayName:
                    type: string
                  email:
                    type: string
        '410':
          description: This link is no longer valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileErrorResponseDto'
      summary: Look up a password-reset token's associated account.
      tags: *ref_8
  /profile/reset-password/token/{token}/confirm:
    post:
      operationId: ProfileController_confirmPasswordReset
      parameters:
        - name: token
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionUserDto'
        '400':
          description: Password must be between 8 and 200 characters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileErrorResponseDto'
        '410':
          description: This link is no longer valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileErrorResponseDto'
      summary: Confirm a password reset via its emailed token, starting a session.
      tags: *ref_8
  /profile/account:
    delete:
      operationId: ProfileController_deleteAccount
      parameters: []
      responses:
        '204':
          description: Deleted.
        '409':
          description: Would leave no active administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileErrorResponseDto'
      security:
        - vaultfolio_session: []
      summary: Delete the caller's own account.
      tags: *ref_8
...
info:
  title: Vaultfolio API
  description: >-
    Generated directly from the backend controllers/DTOs — see
    specs/031-openapi-swagger-integration. "Try it out" reuses the
    vaultfolio_session session cookie from a normal app login; it is not a
    separate credential scheme.
  version: 0.0.9
  contact: {}
...
servers: []
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: vaultfolio_session
  schemas:
...
    SessionUserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          example: admin@example.com
        displayName:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - MEMBER
        domainScopes:
          description: Domain ids this user is entitled to, independent of `role`.
          type: array
          items:
            type: string
      required:
        - id
        - email
        - displayName
        - role
        - domainScopes
...
    ProfileSummaryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - MEMBER
        pendingEmail:
          type: object
          nullable: true
        emailLanguage:
          type: string
          enum: &ref_16
            - en
            - de
          nullable: true
          description: null = not explicitly set; falls back to the default language.
      required:
        - id
        - email
        - displayName
        - role
    UpdateDisplayNameRequestDto:
      type: object
      properties:
        displayName:
          type: string
      required:
        - displayName
    ProfileErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          enum:
            - invalid_display_name
            - email_unavailable
            - email_delivery_failed
            - invalid_token
            - invalid_password
            - invalid_current_password
            - last_admin
            - deletion_failed
            - invalid_email_language
            - bot_protection_failed
        message:
          type: string
      required:
        - error
        - message
    UpdateEmailLanguageRequestDto:
      type: object
      properties:
        emailLanguage:
          type: string
          enum: *ref_16
          nullable: true
      required:
        - emailLanguage
    RequestEmailChangeRequestDto:
      type: object
      properties:
        newEmail:
          type: string
      required:
        - newEmail
    ChangePasswordRequestDto:
      type: object
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
      required:
        - currentPassword
        - newPassword
    ForgotPasswordRequestDto:
      type: object
      properties:
        email:
          type: string
        turnstileToken:
          type: string
          description: Cloudflare Turnstile challenge token.
      required:
        - email
        - turnstileToken
    ResetPasswordRequestDto:
      type: object
      properties:
        newPassword:
          type: string
          format: password
      required:
        - newPassword
...